Reuse existing code for firing events when Element is fully visible.
