+++ client/package.json
@@ -173,7 +173,6 @@
 		"react-day-picker": "^7.4.10",
 		"react-dom": "^17.0.2",
 		"react-element-to-jsx-string": "^14.3.2",
-		"react-intersection-observer": "^8.32.1",
 		"react-live": "^2.3.0",
 		"react-modal": "^3.14.3",
 		"react-query": "^3.32.1",
