+++ client/sites-dashboard/components/sites-table-row.tsx
@@ -3,11 +3,11 @@ import { useSiteLaunchStatusLabel } from '@automattic/sites';
 import { css } from '@emotion/css';
 import styled from '@emotion/styled';
 import { useI18n } from '@wordpress/react-i18n';
-import { memo } from 'react';
-import { useInView } from 'react-intersection-observer';
+import { memo, useState } from 'react';
 import StatsSparkline from 'calypso/blocks/stats-sparkline';
 import JetpackLogo from 'calypso/components/jetpack-logo';
 import TimeSince from 'calypso/components/time-since';
+import { useInView } from 'calypso/lib/use-in-view/index';
 import { displaySiteUrl, getDashboardUrl, isNotAtomicJetpack, MEDIA_QUERIES } from '../utils';
 import { SitesEllipsisMenu } from './sites-ellipsis-menu';
 import SitesP2Badge from './sites-p2-badge';
@@ -81,7 +81,8 @@ const SitePlanIcon = styled.div`
 export default memo( function SitesTableRow( { site }: SiteTableRowProps ) {
 	const { __ } = useI18n();
 	const translatedStatus = useSiteLaunchStatusLabel( site );
-	const { ref, inView: inViewOnce } = useInView( { triggerOnce: true } );
+	const [ hasViewed, setHasViewed ] = useState( false );
+	const ref = useInView< HTMLTableDataCellElement >( () => setHasViewed( true ) );
 
 	const isP2Site = site.options?.is_wpforteams_site;
 
@@ -140,7 +141,7 @@ export default memo( function SitesTableRow( { site }: SiteTableRowProps ) {
 				{ site.options?.updated_at ? <TimeSince date={ site.options.updated_at } /> : '' }
 			</Column>
 			<Column ref={ ref } mobileHidden>
-				{ inViewOnce && (
+				{ hasViewed && (
 					<a href={ `/stats/day/${ site.slug }` }>
 						<StatsSparkline siteId={ site.ID } showLoader={ true }></StatsSparkline>
 					</a>
